DONE
Hopefully this finishes this PR saga.

Metadata review tests are now passing.

The problem was that there were circular dependencies, that weren't
detected with webpack.
