+++ src/app/components/harvest/screens/metadata-review/metadata-review.component.spec.ts
@@ -8,7 +8,7 @@ import { StatisticsComponent } from "@components/harvest/components/shared/stati
 import { Harvest, HarvestMapping, HarvestStatus } from "@models/Harvest";
 import { Project } from "@models/Project";
 import { NgbModal, NgbModalConfig, NgbTooltip } from "@ng-bootstrap/ng-bootstrap";
-import { createRoutingFactory, mockProvider, SpectatorRouting, SpyObject, } from "@ngneat/spectator";
+import { createRoutingFactory, SpectatorRouting, SpyObject, } from "@ngneat/spectator";
 import { generateHarvest } from "@test/fakes/Harvest";
 import { generateProject, generateProjectMeta } from "@test/fakes/Project";
 import { StatisticGroupComponent } from "@components/harvest/components/shared/statistics/group.component";
@@ -25,6 +25,8 @@ import { IconsModule } from "@shared/icons/icons.module";
 import { StrongRouteDirective } from "@directives/strongRoute/strong-route.directive";
 import { provideMockBawApi } from "@baw-api/provide-baw-ApiMock";
 import { NgStyle, DecimalPipe } from "@angular/common";
+import { ToastService } from "@services/toasts/toasts.service";
+import { MockProvider } from "ng-mocks";
 import { FileRowComponent } from "../../components/metadata-review/file-row.component";
 import { LoadMoreComponent } from "../../components/metadata-review/load-more.component";
 import { MetadataReviewComponent } from "./metadata-review.component";
@@ -42,7 +44,7 @@ describe("MetadataReviewComponent", () => {
     component: MetadataReviewComponent,
     providers: [
       provideMockBawApi(),
-      mockProvider(HarvestStagesService, {
+      MockProvider(HarvestStagesService, {
         project: defaultProject,
         harvest: defaultHarvest,
         transition: (_stage: HarvestStatus) => {}
@@ -66,7 +68,7 @@ describe("MetadataReviewComponent", () => {
       SiteSelectorComponent,
       UTCOffsetSelectorComponent,
     ],
-    // mocks: [ToastService],
+    mocks: [ToastService],
   });
 
   function setup(): SpyObject<HarvestStagesService> {
@@ -148,10 +150,10 @@ describe("MetadataReviewComponent", () => {
   afterEach(() => {
     // dismiss all bootstrap modals, so if a test fails
     // it doesn't impact future tests by using a stale modal
-    modalService?.dismissAll();
+    modalService.dismissAll();
   });
 
-  fit("should create", () => {
+  it("should create", () => {
     setup();
     expect(spec.component).toBeInstanceOf(MetadataReviewComponent);
   });
